backend/requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.14,<3.15
djangorestframework-simplejwt>=5.3,<5.4
django-cors-headers>=4.3,<4.4
django-environ>=0.11,<0.12
psycopg2-binary>=2.9,<2.10
gunicorn>=21.2,<21.3
whitenoise>=6.6,<6.7
